Add formatter via spec file rather than command line argument.
The logic to work out which formatters are active is complex, and the hack was
becoming fragile. This should be much more robust.

We now add the formatter by appending a spec file which adds the formatter via
RSpec.configure. Since this file is loaded last, the formatter list should be
finalized, and our logic is simple.